Final commit (refactor/prettier)
