+++ src/test/java/faang/school/postservice/controller/post/PostControllerTest.java
@@ -1,7 +1,7 @@
 package faang.school.postservice.controller.post;
 
 import faang.school.postservice.dto.post.PostDto;
-import faang.school.postservice.service.post.PostService;
+import faang.school.postservice.service.post.PostServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -20,7 +20,7 @@ public class PostControllerTest {
     private PostController postController;
 
     @Mock
-    private PostService postService;
+    private PostServiceImpl postService;
 
     private PostDto postDto;
